Biological Stressor-Response Gradient Lines of Evidence

Use linear regression to evaluate stressor-response gradients inside the case or outside the case.

getBioStressorResponses(
  TargetSiteID,
  df_stressinfo,
  df_respinfo,
  df_respdata,
  df_datapaired,
  biocomm,
  bioindex,
  min_cases = samplim,
  p.val_cutoff = 0.05,
  r2_cutoff = 0.1,
  plotvars = data_plotvars,
  refOutline = refOutline_col,
  plotdpi,
  plotH,
  plotW,
  plotunits,
  dir_plots,
  dir_sub = "_WoE",
  boo_pred_warn = TRUE,
  boo_plot = TRUE
)

Arguments

TargetSiteID

Site ID

df_stressinfo

stressor metadata for stressors considered candidate causes of impairment at the target site

df_respinfo

Bio metric metadata

df_respdata

Biological metric data for all response samples

df_datapaired

dataframe of matched biological response and stressor data.

biocomm

Biological community; algae or BMI. Default = "BMI".

bioindex

Name of the biological index column

min_cases

Minimum number of paired samples; samplim from CASTool_Metadata Default = samplim.

p.val_cutoff

p-value cutoff above which the slope is not considered significant Default = 0.05

r2_cutoff

r2 value below which the relationship has too much variance Default = 0.1

plotvars

colors, shapes, fills, and transparencies for each type (target, not degraded, degraded, inside-the-case, and outside-the-case)

refOutline

color of the reference sites outline

plotdpi

standardized dpi for all plots

plotH

standardized height for all plots

plotW

standardized width for all plots

plotunits

stardardized units for all plots

dir_plots

Directory to save plots. Default = working directory and Results.

dir_sub

Subdirectory for outputs from this function. Default = "StressorResponse"

boo_pred_warn

Should warnings for prediction be suppressed. Default = TRUE.

boo_plot

Boolean value to save plots. Default = TRUE.

Value

Writes one or more graphics depicting stressor-response relationships, a correlation tile plot, and two tab-delimited text files; stressor correlations and scores. Returns a dataframe containing target site scores for inside-the-case and outside-the-case.

Details

Biological (BMI, Algae, or Fish) stressor linear regressions.

Requires packages dplyr, ggplot2, stringr, tidyr

Examples

# None at this time