Co-Occurrence and Sufficiency Lines of Evidence

Generates box plots and and logistic regressions to answer the questions: 1) are the observed stressor levels consistent with impairment where and when it occurs? and 2) are stressor levels sufficient to explain the observed impairment? Also writes a tab-delimited text file containing scores for these two lines of evidence.

getCoOccur(
  TargetSiteID,
  df_data,
  detects,
  df_stressinfo,
  compsites,
  biocomm,
  colBio,
  pHlimLow = 5,
  pHlimHigh = 9,
  DOlim = 6,
  plotvars = data_plotvars,
  plotdpi = plot_dpi,
  plotH = plot_H,
  plotW = plot_W,
  plotunits = plot_units,
  dir_plots = dir_results,
  dir_sub = "_WoE",
  boo_plot = TRUE
)

Arguments

TargetSiteID

ID of station/sample(s) to plot

df_data

data frame with paired stressor/response data, where stressor data are transformed.

detects

all stressors detected in any samples from the target site

df_stressinfo

dataframe containing stressor metadata (e.g., Label, DirIncStress, and other columns from the stressor information metadata file).

compsites

vector of comparator sites. Defaults to list.CompSites$comp.sites

biocomm

Biological community; fish, algae, or BMI. Default = "BMI".

colBio

df_data column with biological index numeric value.

pHlimLow

The lower limit of pH considered to be supportive of a biological community. Defaults to pH of 6.5.

pHlimHigh

The upper limit of pH considered to be supportive of a biological community. Defaults to pH of 9.

DOlim

The lower limit of DO in mg/L considered to be supportive of a biological community. Defaults to 7 mg/L

plotvars

Colors, fills, shapes, transparencies for each type (target, not degraded, degraded, inside-the-case, outside-the-case). Default = data_plotvars.

plotdpi

DPI for plots for standardization. Default = plot_dpi.

plotH

Plot height for standardization. Default = plot_H.

plotW

Plot width for standardization. Default = plot_W.

plotunits

Plot units for standardization. Default = plot_units.

dir_plots

Directory to save plots. Default = working directory and Results.

dir_sub

Subdirectory for outputs from this function. Default = "CoOccurrence"

boo_plot

Boolean value to save plots. Default = TRUE.

Value

Writes individual plots as pngs, and a tab-delimited text file with scores for each line of evidence (co-occurrence & sufficiency) to a "Results/TargetSiteID/BioComm/CoOccurrence" directory. Returns a list containing 3 elements: a dataframe of stressor metadata corresponding with stressors continuing forward in the analysis, a vector of stressors not continuing forward in the analysis, and a dataframe of all the scores continuing forward.

Details

Derive evidence for spatial/temporal co-occurrence.

Stressor-response from field observational studies: Are higher levels of the stressor observed where and when the biological effect occurs?

Box plots are used to show the distribution of the stressor levels at comparator sites with better biological condition. If a site has multiple biological condition scores the lowest score is used to determine "better" sites.

Samples are scored:

1. Supports the case for candidate cause. Stressor levels at the test sites are above the 75th percentile of comparator sites having higher biological quality.

0. Indeterminate. Stressor levels at the test site are below the 50th percentile of comparator sites having higher biological quality.

-1. Weakens the case for the candidate cause. Stressor levels at the test sites are between the 50th and 75th percentile of comparator sites having higher biological quality.

Multiple stressors can be used. Stressors for which all target samples are scored -1 are not included in further lines of evidence evaluation.

Uses the libraries dplyr, ggplot2, grid, stringr.

Examples

# None at this time