Get Matching Stressor-Response Data

Prepare matched stressor/response data as a single dataframe

getCoOccurDataset(
  df_sites,
  df_stress,
  biocomm,
  df_resp,
  index,
  lagdays = c(0, 0)
)

Arguments

df_sites

Dataframe containing site data for all sites.

df_stress

Dataframe containing measured or modeled stressor data for all sites that have them.

biocomm

Biological community; algae or BMI.

df_resp

Dataframe containing biological response metrics.

index

Name of the response index column.

lagdays

c(before, after). The number of days allowed between the stressor sample date and the response sample date, before and after. Default = c(0, 0) (same day).

Value

A dataframe containing matched stressor response data based on a stressor sample obtained between lagdays before and lagdays after the response sample was obtained. Result values are *not* transformed.

Details

Generates a dataframe and writes a table with stressor and response data for all stressor and response samples obtained within a specified time of each other. User specified lag time between stressor sample (prior) and response sample (after) may be employed, with a default of same day samples.

Improvements: Add BioDegLab and BioNarLab?

Uses the libraries dplyr, fuzzyjoin, and tidyr.

Examples

# None at this time