Get Comparator Sites

Generates a list of comparator sites for the target site.

getComparators(
  TargetSiteID,
  df_sites,
  df_cluster,
  df_bioCoOccur,
  bioIndex,
  useBC = FALSE,
  outcaseColName = NULL,
  outcaseLabel = outcaseLabel,
  incaseColName = NULL,
  incaseLabel = incaseLabel,
  useAllCompReaches = FALSE,
  df_bcdist = NULL,
  bc_cutoff = 0.05,
  dir_results,
  dir_sub = "SiteInfo"
)

Arguments

TargetSiteID

Site ID for the site being evaluated

df_sites

Sites table with cluster membership. Default = "data_Sites".

df_cluster

Dataframe containing reach IDs and their associated clusters

df_bioCoOccur

Dataframe containing paired stressor/response sample data

bioIndex

Character vector corresponding to the biological index column name.

useBC

TRUE to use biological similarity; FALSE to not use. Default = "FALSE".

outcaseColName

If useBC == FALSE, all the sites in the sites file (NULL). If useBC == TRUE, the name of the column that indicates the "outside the case". Default = NULL.

outcaseLabel

A label used to describe the "outside the case" region. Examples include "Entire state" or "Omernik Level 3 Ecoregion."

incaseColName

If useBC == FALSE, the name of the column in the sites file that indicates "inside the case". Default = NULL. If NULL, df_bcdist and bc_cutoff must be defined.

incaseLabel

A label used to describe the "inside the case" region. Examples include "Cluster" or "Sites with expected biologic similarity from cluster x" where x refers to the cluster membership of the target site.

useAllCompReaches

Whether all comparator reaches should be used (FALSE) or only those with sites on them (TRUE).

df_bcdist

Dataframe containing the biological dissimilarity distance matrix. Default = NULL. Must be defined if incaseColName is not.

bc_cutoff

Cutoff value below which will be considered similar to the target site. Default = "0.05".

dir_results

Directory containing all results. Default = "file.path(getwd(),"Results")"

dir_sub

Subdirectory for outputs from this function. Default = "SiteInfo"

Value

A list containing 1) a vector of comparator site identifiers (comp.sites), 2) a vector of all sites representing "outside the case", and 3) a dataframe containing gap details (function name, condition, result, and comment); and a tab-delimited text file containing the list of comparator sites with additional information relating to their similarity to the target site.

Details

Obtains the cluster members for the cluster containing the target site and, if useBC = TRUE (default is FALSE), subsets the cluster based on a biological dissimilarity distance matrix (Bray-Curtis) at the specified level of dissimilarity.

Uses the library dplyr, stringr