Report Generation

Generate report

getReport(
  TargetSiteID,
  biocomms,
  primeIndex = bmiIndexGp,
  removeOutliers,
  samplim,
  r2_cutoff,
  p.val_cutoff,
  useBC,
  lagdays,
  DOlim,
  pHlimLow,
  pHlimHigh,
  bmiIndex,
  algIndex,
  fishIndex,
  useBMI,
  useAlg,
  useFish,
  dir_data = normalizePath(file.path(".", "Data")),
  dir_results = normalizePath(file.path(".", "Results")),
  report_type = "full",
  report_format = "html",
  dir_rmd = file.path(system.file(package = "CASTfxn"), "rmd")
)

Arguments

TargetSiteID

SiteID

biocomms

x

primeIndex

x

removeOutliers

Default = TRUE

samplim

x

r2_cutoff

x

p.val_cutoff

x

useBC

Default = TRUE

lagdays

Default = 10, 10

DOlim

x

pHlimLow

x

pHlimHigh

x

bmiIndex

Default = bmiIndex

algIndex

Default = algIndex

fishIndex

Default = fishIndex

useBMI

boolean to use Benthic Macroinvertebrates.

useAlg

boolean to use Algae.

useFish

boolean to use Fish

dir_data

Absolute path to data. Default = /Data

dir_results

Absoluthe path with subfolders named by SiteID. Default = Results folder in working directory.

report_type

Requested report type (all or summary). Default = summary

report_format

Requested report output format (html or word). Default = html

dir_rmd

Directory with template RMD for report. Default = package rmd folder.

Value

A report for the provided SiteID in the provided format (html or word) in the results directory.

Details

Generates a report based on a Target SiteID in a given directory. The output from other `CASTfxn` functions are stored in a "Results" folder with the provided Target SiteID as a subdirectory. Report format can be "html" (recommended default) or "word" (docx).

Only the "summary" report is active.

dir_data and dir_results should be absolute and not relative paths. This is because these directories are passed to an RMD file. Within the RMD each code chunk assumes a relative path based on the RMD location and not the working directory of the calling function. The function `normalizePath` can be used to convert from relative to absolute path.

Examples

TargetSiteID <- "SRCKN001.61"
dir_results <- file.path(getwd(), "Results")
report_type <- "summary"
report_format <- "html"

if (FALSE) { # \dontrun{
# Run Function
getReport(TargetSiteID, dir_results, report_type, report_format)
} # }