Site Info

Get site info of provided siteID.

getSiteInfo(TargetSiteID, dir_results = file.path(getwd(), "Results"),
  data.Stations.Info, data.SampSummary, data.303d.ComID, data.bmi.metrics,
  data.algae.metrics, data.cluster, data.mod, map_proj = NULL,
  map_outline = NULL, map_flowline = NULL, map_flowline2 = NULL,
  dir_sub = "SiteInfo")

Arguments

TargetSiteID
SiteID

dir_results
Directory for results. Default = "Results".

data.Stations.Info
data.Stations.Info

data.SampSummary
data.SampSummary

data.303d.ComID
data.303d.ComID

data.bmi.metrics
data.bmi.metrics

data.algae.metrics
data.algae.metrics

data.cluster
data.cluster

data.mod
data.mod

map_proj
Map projection. If no projection is provided an unprojected map is created without flowlines.

map_outline
Outline for map, typically State border.

map_flowline
Typically NHD+ flowline.

map_flowline2
Typically NHD+ flowline. Can be more than one but plotted the same.

dir_sub
Subdirectory for outputs from this function. Default = "SiteInfo"

Value

A jpg map to a subdirectory "SiteInfo" in the folder named by the SiteID in the user supplied dir_results folder (default is "Results" folder in the working directory). Also produced is a summary list; SiteInfo, Samps, BMImetrics, AlgMetrics, ReachInfo, COMID, ClustIDs, impair, and mods.

Details

Summary info including lat/long, ref status, cluster membership, samps from site, maps

Requires package rgdal.

Required objects:

* data.Stations.Info; StationID_Master, FinalLatitude, FinalLongitude , WaterbodyName, GIS_County, CARefSite_2017, COMID_NHD2

* data.SampSummary; StationID_Master, CollDate, ChemSampleID, PhabSampID , BMI.Metrics.SampID, Algae.Metrics.SampID

* data.303d.ComID; ComID, WATER.BODY.NAME, POLLUTANT, FINAL.LISTING.DECISION

* data.bmi.metrics; StationID_Master, CollDate, CSCI, O_E, MMI_Score

* data.algae.metrics; StationCode, SampleDate, H20, D18, S2

* data.cluster; COMID, H6_noland, H6_land, ElevWs, WsAreaSqKm, PrecipWs, TmeanWs

* data.mod; COMID, ReachModStatus, ModReason

Will create output folder dir_results if it doesn't already exist. The default is "Results". A subdirectory is created for each SiteID.

Examples

TargetSiteID <- "SRCKN001.61"
dir_results <- file.path(getwd(), "Results")

# Data
# data import, example
#data.Stations.Info <- read.delim(paste(myDir.Data,"data.Stations.Info.tab",sep=""))
#data.SampSummary <- read.delim(paste(myDir.Data,"data.SampSummary.tab",sep="")
#                               , na.strings = c(""," "))
#data.303d.ComID <- readRDS(paste0(myDir.Data,"data.303dcomid.RDS"))
#data.bmi.metrics <- read.delim(paste(myDir.Data,"data.bmi.metrics.tab",sep=""))
#data.algae.metrics <- read.delim(paste(myDir.Data,"data.algae.metrics.tab",sep=""))
#data.cluster <- read.delim(paste(myDir.Data,"data.all.clust.tab",sep=""))
#data.mod <- read.delim(paste(myDir.Data,"data.ModPerStatus.tab",sep=""))

# Data getSiteInfo
# data, example included with package
data.Stations.Info <- data_Sites
data.SampSummary   <- data_SampSummary
data.303d.ComID    <- data_303d
data.bmi.metrics   <- data_BMIMetrics
data.algae.metrics <- data_AlgMetrics
data.mod           <- data_ReachMod

#' # Cluster based on elevation category  # need for getSiteInfo and getChemDataSubsets
elev_cat <- toupper(data.Stations.Info[data.Stations.Info[,"StationID_Master"]==TargetSiteID
                    , "ElevCategory"])
if(elev_cat=="HI"){
   data.cluster <- data_Cluster_Hi
} else if(elev_cat=="LO") {
   data.cluster <- data_Cluster_Lo
}

# Map data
# San Diego
#flowline <- rgdal::readOGR(dsn = "data_gis/NHDv2_Flowline_Ecoreg85", layer = "NHDv2_eco85_Project")
#outline <- rgdal::readOGR(dsn = "data_gis/Eco85", layer = "Ecoregion85")
# AZ
map_flowline  <- data_GIS_Flow_HI
map_flowline2 <- data_GIS_Flow_LO
if(elev_cat=="HI"){
   map_flowline <- data_GIS_Flow_HI
} else if(elev_cat=="LO") {
   map_flowline <- data_GIS_Flow_LO
}
map_outline   <- data_GIS_AZ_Outline
# Project site data to USGS Albers Equal Area
usgs.aea <- "+proj=aea +lat_1=29.5 +lat_2=45.5 +lat_0=23
              +lon_0=-96 +x_0=0 +y_0=0 +datum=NAD83
              +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
# projection for outline
my.aea <- "+proj=aea +lat_1=20 +lat_2=60 +lat_0=40 +lon_0=-96 +x_0=0 +y_0=0 
           +datum=NAD83 +units=m +no_defs +ellps=GRS80 +towgs84=0,0,0"
map_proj <- my.aea
# 
dir_sub <- "SiteInfo"

# Run getSiteInfo
list.SiteSummary <- getSiteInfo(TargetSiteID
                                , dir_results
                                , data.Stations.Info
                                , data.SampSummary
                                , data.303d.ComID
                                , data.bmi.metrics
                                , data.algae.metrics
                                , data.cluster
                                , data.mod
                                , map_proj
                                , map_outline
                                , map_flowline
                                , dir_sub=dir_sub)