Stressor List

Get stressor list.

getStressorList(
  TargetSiteID,
  outcaseLabel,
  outcaseID,
  outcaseSites,
  incaseSites,
  refSites,
  siteChem,
  df_Stress,
  chemInfo,
  samplim = 10,
  probsHigh = 0.75,
  probsLow = 0.25,
  DOlim = 7,
  pHlimLow = 6.5,
  pHlimHigh = 9,
  biocommlist,
  listbioParamsDEL,
  plotVars,
  dir_results = file.path(getwd(), "Results"),
  dir_sub = "CandidateCauses"
)

Arguments

TargetSiteID

Site ID

outcaseLabel

Label for the "outside the case" identifier

outcaseID

Name of column in sites file that indicates the "outside the case" id

outcaseSites

Vector containing "outside the case" site IDs

incaseSites

Vector containing "inside the case" site IDs (comparators)

refSites

all reference sites

siteChem

dataframe containing any detected stressor data from target site samples at any time

df_Stress

dataframe containing stressor data

chemInfo

dataframe containing stressor metadata, specifically "Label", "DirIncStress", and "LogTransformYN"

samplim

minimum number of samples required to id stressors as candidate causes. Defaults to 10.

probsHigh

probabilities, high. Defaults to 0.75. For stressors considered to be increasers (stress increases with increasing concentration), values above this are considered to be candidate causes.

probsLow

probabilities, low, Defaults to 0.25.For stressors considered to be decreasers (stress increases with decreasing concentration), values above this are considered to be candidate causes.

DOlim

Dissolved Oxygen limit, default = 7. Used to exclude values < probsLow if they are also > 7 mg/L.

pHlimLow

pH limit, low, default = 6.5. Used to exclude values < probsLow if they are also between 6.5 and 9 pH units.

pHlimHigh

pH limit, high, default = 9. Used to exclude values > probsHigh if they are also between 6.5 and 9 pH units.

biocommlist

vector of each biological response community available ("bmi", "alg", or "fish")

listbioParamsDEL

list of vectors corresponding to biocommlist of stressors not considered relevant for evaluation

plotVars

x

dir_results

Directory to save plots. Default = file.path(getwd(), "Results").

dir_sub

Subdirectory for outputs from this function. Default = "SiteInfo"

Value

One or more png files containing normalized stressor boxplots by stressor group; a correlation matrix representing stressor correlations; a file of stressors evaluated; a file of stressors excluded; stressor values and site.stressor.pctrank.

Details

Box plots of each stressor, grouped by category.

Required objects: all specified as inputs.

Required packages: dplyr, ggplot2, stringr, tidyr

chem.info need to include DirIncStress. Valid values are 'inc' or 'dec'.

Examples

# None at this time