Stressor Sufficiency Line of Evidence

Generates logistic regressions to answer the question: are stressor levels sufficient to explain the observed impairment? Also writes a tab-delimited text file containing scores for this line of evidence.

getSufficiency(
  TargetSiteID,
  df_data,
  compSites,
  df_stressinfo,
  biocomm,
  colBio,
  plotvars = data_plotvars,
  plotdpi = plot_dpi,
  plotH = plot_H,
  plotW = plot_W,
  plotunits = plot_units,
  dir_plots = file.path(getwd(), "Results"),
  dir_sub = "_WoE",
  boo_plot = TRUE
)

Arguments

TargetSiteID

ID of station to be evaluated. May have one or many samples.

df_data

dataframe containing matched stressor-response data for the biological response community desired

compSites

vector containing comparator site IDs

df_stressinfo

dataframe containing stressor metadata (LogTransf, Label)

biocomm

Biological community; BMI, algae, or fish Default = "BMI".

colBio

df_data column name for the field with biological index value.

plotvars

Standardized sizes, fills, shapes, and transparencies for plots. Defaults to data_plotvars

plotdpi

Standardized plot dpi. Defaults to plot_dpi.

plotH

Standardized plot height. Defaults to plot_H.

plotW

Standardized plot width. Defaults to plot_W.

plotunits

Standardized plot units. Defaults to plot_units.

dir_plots

Directory to save plots. Default = working directory and Results.

dir_sub

Subdirectory for outputs from this function. Default = "Sufficiency"

boo_plot

Boolean value indicating whether or not to print the plot. Defaults to TRUE.

Value

Writes individual plots as pngs, and a tab-delimited text file with scores for the sufficiency line of evidence to a directory: "Results/TargetSiteID/BioComm/Sufficiency".

Details

Derive Evidence for Stressor Sufficiency from Field Observational Studies.

Stressor-response from field observational studies: Is the level of the stressor sufficient to explain the level of biological effect observed at the site?

Using all comparator sites, fit logistical regression curve of the probability of poor condition (i.e., poor biological index score) as a function of stressor level. Compare stressor levels from test site to levels corresponding to median (50 condition.

1. Supports the case for the candidate cause. Stressor levels at the test site are above the lower confidence limit (LCL) corresponding to 50 probability of observing poor condition

0. Indeterminate. Stressor levels at the test site are between the LCL corresponding to 50 corresponding to 20

-1. Weakens the case for the candidate cause. Stressor levels at the test site are below the upper confidence limit (UCL) corresponding to 20 probability of observing poor condition.

Cut function is used to assign narrative categories and degraded status based on the provided biological score, ensuring criteria are applied the same across all sites.

The BioDegLab has to remain as the default values of Yes and No. Other values will break the code.

Only a single biological measurement is used. But multiple stressors can be used.

Uses the libraries dplyr, ggplot2, grid, stringr, and tidyr.

Examples

# None at this time