Time Sequence Line of Evidence (Graphics only)

Graph time-specific stressor-response values.

getTimeSeq(
  TargetSiteID,
  biocomm,
  bioindex,
  df_stress,
  df_resp,
  df_respinfo,
  df_stressinfo,
  df_paired,
  plotdpi = 600,
  plotH = 6,
  plotW = 8,
  plotunits = "in",
  dir_results = file.path(getwd(), "Results"),
  dir_sub = "_WoE",
  boo_plot = TRUE
)

Arguments

TargetSiteID

site identifier for the site being evaluated (the Target Site)

biocomm

biological response community (bmi, algae, or fish)

bioindex

the column name containing the biological index value

df_stress

dataframe containing all target site stressors, regardless of matching status.

df_resp

dataframe containing the target site biological response samples' index and metric values, regardless of matching status

df_respinfo

dataframe containing response metadata, specifically "Label"

df_stressinfo

dataframe containing stressor metadata, specifically "Label"

df_paired

x

plotdpi

Standardized dpi value for graphics. Default = plot_dpi. 600

plotH

Standardized height value for graphics. Default = plot_H. 6

plotW

Standardized width value for graphics. Default = plot_W. 8

plotunits

Standardized units for graphics. Default = plot_units. "in"

dir_results

Directory containing all results. Default = file.path(getwd(),"Results"). NOTE: the code adds a middle directory, BioComm, between the Results and the subdirectory.

dir_sub

Subdirectory for outputs from this function. Default = "TimeSequence"

boo_plot

Flag declaring whether the plots should be saved or not. Default = TRUE.

Value

One or more pngs in SiteID/Biocomm/TemporalSequence subfolder of the "Results" folder. No scores are currently generated.

Details

Generates faceted time sequence graphics (stressor/response, one atop the other). All stressor/response data are graphed. Improvements: Add scoring.

Uses the packages dplyr, ggplot2, lubridate, tidyr

Examples

# None at this time