Weight-of-Evidence summary

Summarize weight of evidence using scores from other functions.

getWoE(
  TargetSiteID,
  biocomm,
  dfLoE,
  dfStress = df_stressorMetadata,
  dir_results = file.path(getwd(), "Results"),
  dir_WoE = "_WoE"
)

Arguments

TargetSiteID

Site ID

biocomm

Biological community; algae or BMI. Default = "BMI".

dfLoE

Dataframe containing the lines of evidence scores for each stressor and sample in long form.

dir_results

Directory to save tables. Default = working directory and Results.

Value

Two tab-delimited tables containing weight of evidence information: one detailed, and one summary. The detailed table includes stressors and the lines of evidence for or against them. The summary table includes the number of lines that support, refute, are equivalent, or were not evaluated.

Details

Stressor-based weight of evidence for the stressor as a cause of impairment for the specified biological community. Combines information from the co-occurrence, stressor-response using data from or outside the case, verified predictions, and stressor-response from laboratory data.

Uses the packages dplyr and tidyr.