prepSiteData

Prepare measured stressor data for CASTool

prepModStressorData(
  in.dir,
  out.dir,
  fn.data,
  fn.meta,
  removeOutliers,
  sub.dir = "_Histograms"
)

Arguments

in.dir

x

out.dir

directory where all rds files are stored

fn.data

x

fn.meta

x

removeOutliers

x

sub.dir

x

Value

A list containing data_Sites and data_cluster to be used in the CASTool.

Details

Reviews each uploaded file against both user-input data (either via the shiny app or contained in the CASToolMetadata.xlsx file) to evaluate 1) if the file exists, 2) if it contains the required columns (which requires that they be named correctly), 3) whether datatypes meet requirements.

If any required files do not exist, return the list of missing files and shut down. If all required files exist, proceed to internal checks of the columns and data. Lastly, perform minimal joins to determine missing values.

dir_data and dir_results should be absolute and not relative paths. The function `normalizePath` can be used to convert from relative to absolute path.