Site Info

Get site info of provided siteID.

getSiteInfo(
  TargetSiteID,
  TargetCOMID,
  df_Sites,
  df_SampSummary,
  biocommlist,
  df_BMIMetrics = NULL,
  BMIIndexGp = NULL,
  df_ALGMetrics = NULL,
  ALGIndexGp = NULL,
  df_FishMetrics = NULL,
  FishIndexGp = NULL,
  comp.sites,
  all.sites,
  IncaseLabel = NULL,
  OutcaseLabel = NULL,
  useBC = FALSE,
  plotvars = data_plotvars,
  refSiteCol = refOutline_col,
  plotdpi = plot_dpi,
  plotH = plot_H,
  plotW = plot_W,
  plotunits = plot_units,
  dir_photo = file.path(getwd(), "Data", "Photos"),
  dir_results = file.path(getwd(), "Results"),
  dir_sub = "SiteInfo",
  boo_plot = TRUE
)

Arguments

TargetSiteID

site identifier for the site being evaluated (the Target Site)

TargetCOMID

x

df_Sites

dataframe containing site data, including "inside the case" and "outside the case" identifiers. If useBC == TRUE, "outside the case" will be cluster; if useBC == FALSE, "inside the case" will be cluster.

df_SampSummary

dataframe containing sample IDs for samples collected at the target site, organized by sample date (rows) and type (columns)

biocommlist

vector of all biological response communities to be evaluated

df_BMIMetrics

dataframe containing BMI sample index and metric values

BMIIndexGp

vector containing one or more BMI indices for display purpose only

df_ALGMetrics

dataframe containing algae sample index and metric values. Default is NULL.

ALGIndexGp

vector containing one or more algal indices for display purpose only

df_FishMetrics

dataframe containing fish sample index and metric values

FishIndexGp

vector containing one or more fish indices for display purpose only

comp.sites

vector containing inside-the-case (comparator) site IDs

all.sites

vector containing all outside-the-case site IDs

IncaseLabel

Label for the "inside the case" identifier. Default = NULL.

OutcaseLabel

Label for the "outside the case" identifier. Default = NULL.

useBC

TRUE to use biological similarity; FALSE to not use. Default = FALSE.

plotvars

Colors, fills, shapes, transparencies for each type (target, not degraded, degraded, inside-the-case, outside-the-case). Default = data_plotvars.

refSiteCol

Default color outline for reference sites, used for standardization. Default = refOutline_col.

plotdpi

Default dpi for plots, used for standardization. Default = plot_dpi.

plotH

Default height for plots, used for standardization. Default = plot_H.

plotW

Default width for plots, used for standardization. Default = plot_W.

plotunits

Default units for plots, used for standardization. Default = plot_units.

dir_photo

directory containing all site photos (for every site in the data set). Default is file.path(getwd(), "Data", "Photos").

dir_results

Directory containing all results. Default is file.path(getwd(),"Results").

dir_sub

Subdirectory for outputs from this function. Default = "SiteInfo".

boo_plot

Boolean value to save plots. Default = TRUE.

Value

A list containing a dataframe of site information (site ID, geographic coordinates, waterbody name, county name, reach identifier, cluster identifier, and reference site flag); a dataframe of any samples obtained from the site location by sample date; a dataframe of bmi response indices and metric values; a dataframe of algae response indices and metric values; the reach identifier for the target site location; the cluster identifier for the target site; and a vector of reference reaches. Also outputs graphics depicting background variables by category, response indices, and tabular site background data. Creates a subfolder for site photos within the directory "~/Results/TargetSiteID/SiteInfo".

Details

Summary info including lat/long, ref status, cluster membership, samples from site

Requires packages dplyr, ggplot2, lubridate, tidyr

Examples

# None at this time